suggestions/requirements.txt
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf==5.29.6
watchdog==6.0.0

# grpcio==1.78.1
# grpcio-tools==1.78.1
# protobuf==6.33.5
# watchdog==6.0.0

# grpcio==1.60.0
# grpcio-tools==1.60.0
# protobuf==4.25.2
# watchdog==6.0.0